rename files to dm_* change Vector to special version of matrix
