+++ .gitignore
@@ -90,7 +90,15 @@ Temporary Items
 /bazel-*
 /bazel.iml
 /output/
-.bazelversion 
+.bazelversion
 user.bazelrc
 
 /tmp/*
+
+### Added by Hedron's Bazel Compile Commands Extractor: https://github.com/hedronvision/bazel-compile-commands-extractor
+# The external link: Differs on Windows vs macOS/Linux, so we can't check it in. The pattern needs to not have a trailing / because it's a symlink on macOS/Linux.
+/external
+# Compiled output -> don't check in
+/compile_commands.json
+# Directory where clangd puts its indexing work
+/.cache/
