+++ .vscode/c_cpp_properties.json
@@ -8,6 +8,7 @@
                 "${workspaceFolder}/lib/**",
                 "${workspaceFolder}/bazel-03_DoubleMatrix/external/Unity/src",
                 "${workspaceFolder}/bazel-03_DoubleMatrix/external/CException/lib",
+                "${workspaceFolder}/bazel-03_DoubleMatrix/external/dbg-macro/include",
                 "/opt/homebrew/Cellar/glib/2.74.0/**"
             ],
             "defines": [],
