+++ examples/BUILD
@@ -3,8 +3,9 @@ cc_binary(
     srcs = ["example_matrix.c"],
     deps = [
         "//lib/misc",
-        "//src:DoubleMatrix",
-        "//src:DoubleVector",
+        "//src:dm_io",
+        "//src:dm_math",
+        "//src:dm_matrix",
     ],
 )
 
@@ -13,6 +14,8 @@ cc_binary(
     srcs = ["example_vector.c"],
     deps = [
         "//lib/misc",
-        "//src:DoubleVector",
+        "//src:dm_io",
+        "//src:dm_math",
+        "//src:dm_matrix",
     ],
 )
