+++ examples/linear_algebra/BUILD
@@ -3,7 +3,9 @@ cc_binary(
     srcs = ["vec_math.c"],
     deps = [
         "//lib/misc",
-        "//src:DoubleVector",
+        "//src:dm_io",
+        "//src:dm_math",
+        "//src:dm_matrix",
     ],
 )
 
@@ -12,6 +14,8 @@ cc_binary(
     srcs = ["vec_column.c"],
     deps = [
         "//lib/misc",
-        "//src:DoubleVector",
+        "//src:dm_io",
+        "//src:dm_math",
+        "//src:dm_matrix",
     ],
 )
