+++ graph.in
@@ -1,0 +1,21 @@
+digraph mygraph {
+  node [shape=box];
+  "//src:dm_io"
+  "//src:dm_io" -> "//src:include/dm_io.h\n//src:dm_io.c"
+  "//src:dm_io" -> "//src:dm_matrix"
+  "//src:dm_io" -> "//lib/misc:misc"
+  "//src:BUILD"
+  "//src:include/dm_io.h\n//src:dm_io.c"
+  "//src:dm_math"
+  "//src:dm_math" -> "//src:dm_math.c\n//src:include/dm_math.h"
+  "//src:dm_math" -> "//src:dm_matrix"
+  "//src:dm_math" -> "//lib/misc:misc"
+  "//src:dm_matrix"
+  "//src:dm_matrix" -> "//src:include/dm_matrix.h\n//src:dm_matrix.c"
+  "//src:dm_matrix" -> "//lib/misc:misc"
+  "//lib/misc:misc"
+  "//lib/misc:misc" -> "//lib/misc:misc.c\n//lib/misc:misc.h"
+  "//src:include/dm_matrix.h\n//src:dm_matrix.c"
+  "//src:dm_math.c\n//src:include/dm_math.h"
+  "//lib/misc:misc.c\n//lib/misc:misc.h"
+}
