+++ src/include/dm_matrix.h
@@ -1,34 +1,29 @@
-/**
- * @file matrix.h
- * @author Uwe Röttgermann (uwe@roettgermann.de)
- * @brief
- * @version 0.2
- * @date 26-12-2022
- *
- * @copyright Copyright (c) 2022
- *
- */
-
 #ifndef MATRIX_UR_H
 #define MATRIX_UR_H
 
+#include <stdbool.h>
+
 #include "misc.h"
-#include "vector.h"
 
 /*******************************/
 /*     Define & Types         */
 /*******************************/
 
-#define INIT_CAPACITY 5u
-
 // Definition of DoubleMatrix
-typedef struct {
+typedef struct Matrix {
   double **values;
   size_t rows;
   size_t columns;
   size_t column_capacity;
   size_t row_capacity;
 } DoubleMatrix;
+
+// Definition of DoubleVector
+typedef struct Vector {
+  DoubleMatrix *mat1D;
+  size_t length;
+  bool column_vec;
+} DoubleVector;
 
 /*******************************/
 /*        Double Matrix        */
@@ -40,8 +35,7 @@ DoubleMatrix *createDoubleMatrix(size_t rows, size_t cols);
 DoubleMatrix *createRandomDoubleMatrix(size_t rows, size_t cols);
 DoubleMatrix *cloneMatrix(DoubleMatrix *m);
 DoubleMatrix *createIdentityMatrix(size_t rows);
-DoubleMatrix *setArrayToMatrix(size_t rows, size_t cols,
-                               double array[rows][cols]);
+DoubleMatrix *setArrayToMatrix(size_t rows, size_t cols, double **array);
 
 // shrink, push, pop, expand
 void expandMatrixRow(DoubleMatrix *mat);
@@ -50,23 +44,33 @@ void shrinkMatrixColumn(DoubleMatrix *mat);
 void shrinkMatrixRow(DoubleMatrix *mat);
 void pushColumn(DoubleMatrix *mat, DoubleVector *col_vec);
 void pushRow(DoubleMatrix *mat, DoubleVector *row_vec);
+
+// free
+void freeDoubleMatrix(DoubleMatrix *mat);
+
+/*******************************/
+/*  Double Vector  (Dynamic)   */
+/*******************************/
+
+// Function Double Vector:
+DoubleVector *newDoubleVector();
+DoubleVector *newDoubleVectorOfLength(size_t length, double value);
+DoubleVector *newRandomDoubleVectorOfLength(size_t length);
+DoubleVector *cloneDoubleVector(const DoubleVector *vector);
 DoubleVector *popColumn(DoubleMatrix *mat);
 DoubleVector *popRow(DoubleMatrix *mat);
-
-// free & print
-void freeDoubleMatrix(DoubleMatrix *mat);
-void printDoubleMatrix(DoubleMatrix *matrix);
-
-// math:
 DoubleVector *getRowVector(DoubleMatrix *mat, size_t row);
 DoubleVector *getColumnVector(DoubleMatrix *mat, size_t column);
-double *showRowVector(DoubleMatrix *mat, size_t row);
-void scalarMultiply(DoubleMatrix *mat, double scalar);
-void transpose(DoubleMatrix *mat);
-DoubleMatrix *multiplyDoubleMatrices(DoubleMatrix *m1, DoubleMatrix *m2);
-DoubleVector *multiplyVectorMatrix(DoubleVector *vec, DoubleMatrix *mat);
-bool areEqualDoubleMatrices(DoubleMatrix *m1, DoubleMatrix *m2);
+void setDoubleVectorArray(DoubleVector *vec, double *array, size_t len_array);
+double *getArrayFromVector(DoubleVector *vec);
 
-double vector_multiply(double *col, double *row, size_t length);
+// shrink, push, pop, expand
+void expandDoubleVector(DoubleVector *vec);
+void shrinkDoubleVector(DoubleVector *vec);
+void pushValue(DoubleVector *vec, double value);
+double popValue(DoubleVector *vec);
+
+// free
+void freeDoubleVector(DoubleVector *vec);
 
 #endif  // !MATRIX_H
